pipeline/requirements.txt
apache-beam[gcp]==2.12.0
avro==1.9.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
crcmod==1.7
dill==0.2.9
enum34==1.1.6
fastavro==0.21.24
fasteners==0.15
future==0.17.1
futures==3.2.0
google-api-core==1.11.1
google-apitools==0.5.26
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.0
google-cloud-videointelligence==1.9.0
google-cloud-vision==0.36.0
google-resumable-media==0.3.2
google-api-python-client==1.6.5
googleapis-common-protos==1.6.0
grpcio==1.21.1
httplib2==0.11.3
idna==2.8
monotonic==1.5
numpy==1.16.4
oauth2client==3.0.0
protobuf==3.8.0
pyarrow==0.11.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
PyYAML==3.13
requests==2.22.0
rsa==4.0
six==1.12.0
typing==3.6.6
urllib3==1.25.3
